Ensure event listeners unbind when the audio player controls directive is destroyed.
If an album was searched and then added to the queue, it was causing multiple 'audio.ended' events
which was causing the queue to skip every other track. This was due to a duplicate instance of
the audio player controls directive listening to the event (or a duplicate listener on rootScope, not
sure which).

Fix is to manually unbind the event listeners from the $rootScope when $destroy is called