Moved some styles into seperate files for pages.
Also fixed issues with updating scope so that scope is not set until all pending API requests complete.
This fixes the issue whene the directive for the action menus get's an undefined playlist or other entity
because the template was rendered before the requests were all complete. Need to ensure ALL requests finish
so that paramters to directives are set properly