WARNING: Wall of changes
Mainly, this is a rework of the storage from a local disk storage to AWS S3.

The indexer has been reworked to a stand-alone app that scans for tracks and uploads them
to S3 and sends the meta data to the main server.

The concept of a 'library' has been removed. In the future, I may add support for Google storage
or another major storage vendor.

The app has been split into 3 gradle sub projects: 'indexer', 'server' and 'api'.

'api' is shared code between the 'server' and 'indexer'.

Most of the major funcionality is written, but there are some TODO's leftover that will need
to be implemented in order for the app to work properly.